requirements.txt
fastapi
uvicorn
openai
PyMuPDF
sentence-transformers
faiss-cpu
pydantic
python-multipart
streamlit
requests
groq
